finished adding all the memory and cpu boilerplate

+++ lib/core/include/cpu3026/core/types.h
@@ -5,7 +5,7 @@
 #include <cstddef>
 
 namespace cpu3026 {
-	using byte_t = std::uint16_t;
+	using byte_t = std::uint8_t;
 	using word_t = std::uint16_t;
 	using dword_t = std::uint32_t;
 	using address_t = std::uint16_t;
